refactor: move lint logic from pacakger to command
